fix tests after merge
