+++ tests/conftest.py
@@ -233,6 +233,31 @@ def random_kheader(request, random_full_ismrmrd_header, random_acq_info):
     return kheader
 
 
+@pytest.fixture(params=({'seed': 0, 'n_other': 10, 'n_k2': 40, 'n_k1': 20},))
+def random_kheader_shape(request, random_acquisition, random_full_ismrmrd_header):
+    """Random (not necessarily valid) KHeader with defined shape."""
+    # Get dimensions
+    seed, n_other, n_k2, n_k1 = (
+        request.param['seed'],
+        request.param['n_other'],
+        request.param['n_k2'],
+        request.param['n_k1'],
+    )
+    generator = RandomGenerator(seed)
+
+    # Generate acquisitions
+    random_acq_info = AcqInfo.from_ismrmrd_acquisitions([random_acquisition for _ in range(n_k1 * n_k2 * n_other)])
+    n_k0 = int(random_acq_info.number_of_samples[0])
+    n_coils = int(random_acq_info.active_channels[0])
+
+    # Generate trajectory
+    ktraj = [generate_random_trajectory(generator, shape=(n_k0, 2)) for _ in range(n_k1 * n_k2 * n_other)]
+
+    # Put it all together to a KHeader object
+    kheader = KHeader.from_ismrmrd(random_full_ismrmrd_header, acq_info=random_acq_info, defaults={'trajectory': ktraj})
+    return kheader, n_other, n_coils, n_k2, n_k1, n_k0
+
+
 @pytest.fixture(params=({'seed': 0, 'n_other': 2, 'n_coils': 16, 'n_z': 32, 'n_y': 128, 'n_x': 256},))
 def random_test_data(request):
     seed, n_other, n_coils, n_z, n_y, n_x = (
